data-access/requirements.txt
fastapi==0.85.1
uvicorn==0.18.3
psycopg2-binary==2.9.5
